Enabler Tool for HR Application using java full stack
This repository contains both the frontend and backend files for the Enabler Tool project, designed for an HR application. The project was developed using the following technologies:
Backend: Spring Boot
Frontend: JavaScript, HTML/CSS, Bootstrap
Database: MySQL
Project Overview:
The Enabler Tool is specifically designed to fetch and display details of inactive users within the HR system. The application also includes a feature to download a copy of the fetched data for further processing or archival purposes.
Both the frontend and backend components work seamlessly together to provide an efficient and user-friendly interface for HR professionals to manage and track inactive users.